fix(fetch): Request body methods callable as value / computed key (#1698)
`req.json()` worked as a direct typed call (#1688) but failed when accessed
as a value or via a computed key:

  typeof req.json            // was 'object'    → now 'function'
  typeof (req as any)[key]   // was 'undefined' → now 'function'
  await (req as any)[key]()  // was null        → now the parsed body

The value-read / computed-key / `as any` forms lose the static `Request`
type at codegen, so they land in the runtime's generic dynamic-dispatch
tower (`js_native_call_method` → small-handle → `js_handle_method_dispatch`
/ `js_handle_property_dispatch`), which had no Request body-method arm. This
blocked Hono's `c.req.json()` end-to-end (#1655): HonoRequest's `#cachedBody`
does `raw[key]()` — a computed-key call on the underlying Request.

Fix (runtime dispatch, type-agnostic so it covers typed / computed / `as any`
/ Hono's untyped `raw`):

- `dispatch_request_method` (text/json/arrayBuffer) wired into
  `js_handle_method_dispatch` — mirrors the existing `dispatch_response_method`.
- `dispatch_request_property` returns `js_class_method_bind` bound-method
  values for json/text/arrayBuffer, so `typeof req[key] === 'function'` and the
  value stays callable (#1670 stream-property pattern).
- Unified fetch-family handle id counter (`NEXT_FETCH_HANDLE_ID`) across
  Request/Response/Headers/Blob. Previously each had its own counter starting
  at 1, so a Request id could equal a Response id and the registry-membership
  dispatch gate was ambiguous. One counter ⇒ an id belongs to exactly one
  registry ⇒ unambiguous dispatch (also makes `new Response(body)[key]()` and
  Response computed-key/typeof work).
- `typeof req.json` codegen fold for the LITERAL value-read (it takes the typed
  Web-Fetch path that bypasses the property dispatch). Covers Request + Response.

Verified: the issue repro, computed/as-any/typed call+typeof variants, and a
Hono POST `/api/echo` doing `await c.req.json()` all match `node
--experimental-strip-types` byte-for-byte. `test_gap_fetch_response.ts`
(Response/Headers/Request/Blob/redirect) is unchanged. perry-hir + perry-stdlib
unit tests pass.

Note: by-default Hono compile still hits a SEPARATE, pre-existing guard —
`#503` refuses hono's internal `path[runtimeVar]()` dynamic dispatch (untouched
by this change). Tracked separately for #1655.

Test: test-files/test_issue_1698_request_body_method_value.ts
